Updated export to json to not overwrite existing files by default
but provided an override to allow you to force them to be overwritten

+++ src/gatsby-node.js
@@ -12,7 +12,10 @@ const writeNodeToJson = ( post, options ) => {
 		const postDate = new Date( post.publishedOn );
 		const fileTitle = postDate.getFullYear() + "-" + ( "0" + ( postDate.getMonth() + 1 ) ).slice( -2 ) + "-" + ( "0" + postDate.getDate() ).slice( -2 ) + "-" + _.kebabCase( post.slug ) + ".json";
 		const filePath = path.join( options.exportToJson.path, `/${ fileTitle }` );
-		fs.writeFileSync( filePath, JSON.stringify( post, null, 4 ), "utf8" )
+		const fileAlreadyExists = fs.existsSync( filePath );
+		if ( options.exportToJson.overwriteExisting || !fileAlreadyExists ) {
+			fs.writeFileSync( filePath, JSON.stringify( post, null, 4 ), "utf8" )
+		}
 	}
 };
 
@@ -110,7 +113,8 @@ exports.sourceNodes = async ( { boundActionCreators, reporter, store }, pluginOp
 		replaceStrings: [],
 		exportToJson: {
 			enabled: false,
-			path: ""
+			path: "",
+			overwriteExisting: false
 		},
 		query: {
 			categoryId: null,
